Only perform cleanup at most once every 24 hours
This commit writes a `gc.properties` file into the current version's
cache directory and touches it when it has performed the cleanup.
If it's not present or hasn't been touched within the last 24 hours,
the cleanup is performed; otherwise, it is skipped. This is done so we
don't scan version-specific caches more often than it makes sense.

Issue: #5742