Fix resolution error not captured in build operation
This commit fixes the ResolveConfigurationDependenciesBuildOperationType
failure capture, which failed after refactoring the way errors are thrown
during dependency resolution. The change was made in order to make it
possible for tasks to react on resolution errors lazily, rather than
eagerly (see ResolvableDependenciesInternal#getResolutionResult(Action<? super Throwable> errorHandler)).

To achieve this, it is now possible to set a lazy producer on the build
operation context. Then only when we need the result of the execution in the
listener, the result would be produced. If an error is thrown during producing,
then the operation result is deemed failed, otherwise we take the result of
the execution of the deferred operation as the result.